Комментарии, удалил лишнее, разнес по папкам
