+++ TestApp/ParkingApp/ParkingApp.csproj
@@ -23,5 +23,9 @@
         <DependentUpon>ParkingPage.xaml</DependentUpon>
         <SubType>Code</SubType>
       </Compile>
+      <Compile Update="Pages\App.xaml.cs">
+        <DependentUpon>App.xaml</DependentUpon>
+        <SubType>Code</SubType>
+      </Compile>
     </ItemGroup>
 </Project>
