+++ TestApp/TestApp.Android/MainActivity.cs
@@ -6,6 +6,7 @@ using Android.Views;
 using Android.Widget;
 using Android.OS;
 using ParkingApp;
+using ParkingApp.Pages;
 
 namespace TestApp.Android
 {
