+++ TestApp/TestApp.iOS/AppDelegate.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Foundation;
 using ParkingApp;
+using ParkingApp.Pages;
 using UIKit;
 
 namespace TestApp.iOS
